lambdas/agent/requirements.txt
strands-agents>=1.0.0
boto3>=1.34.0
httpx>=0.27.0
